requirements.txt
_openmp_mutex==4.5
aws-c-auth==0.7.31
aws-c-cal==0.7.4
aws-c-common==0.9.28
aws-c-compression==0.2.19
aws-c-event-stream==0.4.3
aws-c-http==0.8.10
aws-c-io==0.14.18
aws-c-mqtt==0.10.6
aws-c-s3==0.6.6
aws-c-sdkutils==0.1.19
aws-checksums==0.1.20
aws-crt-cpp==0.28.3
aws-sdk-cpp==1.11.407
blinker==1.8.2
branca==0.7.2
brotli-python==1.1.0
bzip2==1.0.8
c-ares==1.33.1
ca-certificates==2024.8.30
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
csp==0.0.5
csv23==0.3.4
cyrus-sasl==2.1.27
exprtk==0.0.2
flask==3.0.3
folium==0.17.0
gbfs-client==0.1.8
greenlet==3.1.1
gtest==1.15.2
h2==4.1.0
haversine==2.8.1
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
importlib-metadata==8.5.0
intel-openmp==2024.2.1
itsdangerous==2.2.0
jinja2==3.1.4
krb5==1.21.3
libabseil==20240116.2
libarrow-acero==16.1.0
libarrow-dataset==16.1.0
libarrow-substrait==16.1.0
libarrow==16.1.0
libblas==3.9.0
libbrotlicommon==1.1.0
libbrotlidec==1.1.0
libbrotlienc==1.1.0
libcblas==3.9.0
libcrc32c==1.1.2
libcurl==8.10.1
libdb==6.2.32
libevent==2.1.12
libffi==3.4.2
libgcc==14.1.0
libgomp==14.1.0
libgoogle-cloud-storage==2.29.0
libgoogle-cloud==2.29.0
libgrpc==1.62.2
liblapack==3.9.0
libparquet==16.1.0
libprotobuf==4.25.3
librdkafka==2.4.0
libre2-11==2023.09.01
libsqlite==3.46.1
libssh2==1.11.0
libthrift==0.20.0
libutf8proc==2.8.0
libwinpthread==12.0.0.r4.gg4f2fc60ca
libxcb==1.17.0
libzlib==1.3.1
lz4-c==1.9.4
markupsafe==2.1.5
mkl==2024.1.0
numpy==1.26.4
openssl==3.3.2
orc==2.0.2
packaging==24.1
pandas==2.2.3
pip==24.2
psutil==6.0.0
pthread-stubs==0.4
pyarrow-core==16.1.0
pyarrow==16.1.0
pycparser==2.22
pysocks==1.7.1
python-dateutil==2.9.0
python-tzdata==2024.2
python==3.9.20
python_abi==3.9
pytz==2024.1
rapidjson==1.1.0.post20240409
re2==2023.09.01
requests==2.32.3
ruamel.yaml.clib==0.2.8
ruamel.yaml==0.18.6
setuptools==75.1.0
six==1.16.0
snappy==1.2.1
sqlalchemy==2.0.35
tbb==2021.7.0
tk==8.6.13
typing-extensions==4.12.2
typing_extensions==4.12.2
tzdata==2024a
ucrt==10.0.22621.0
urllib3==2.2.3
vc14_runtime==14.40.33810
vc==14.3
vs2015_runtime==14.40.33810
werkzeug==3.0.4
wheel==0.44.0
win_inet_pton==1.1.0
xorg-libx11==1.8.10
xorg-libxau==1.0.11
xorg-libxdmcp==1.1.3
xorg-xorgproto==2024.1
xyzservices==2024.9.0
xz==5.2.6
zipp==3.20.2
zstandard==0.23.0
zstd==1.5.6
